refactoring auth

+++ project.py
@@ -28,17 +28,10 @@ DBSession = sessionmaker(bind=engine)
 session = DBSession()
 
 
-# Create anti-forgery state token
-@app.route('/login')
-def showLogin():
-    state = ''.join(random.choice(string.ascii_uppercase + string.digits)
-                for x in range(32))
-    login_session['state'] = state
-    # return "The current session state is %s" % login_session['state']
-    return render_template('login.html', STATE=state)
 
 # User Helper Functions
 def createUser(login_session):
+
     newUser = User(name=login_session['username'], email=login_session['email'], picture=login_session['picture'])
     session.add(newUser)
     session.commit()
@@ -57,211 +50,201 @@ def getUserID(email):
         return user.id
     except:
         return None
-###########
-#FBconnect#
-##########
+
+# Create anti-forgery state token
+@app.route('/Catalog/login')
+def login():
+	# Create anti-forgery state token
+	state = ''.join(random.choice(string.ascii_uppercase + string.digits) for x in range(32))
+	login_session['state'] = state
+
+	return render_template('login.html', STATE=state)
+
+@app.route('/Catalog/logout')
+def logout():
+	if login_session['provider'] == 'facebook':
+		fbdisconnect()
+		del login_session['facebook_id']
+
+	if login_session['provider'] == 'google':
+		gdisconnect()
+		del login_session['gplus_id']
+		del login_session['access_token']
+
+	del login_session['username']
+	del login_session['email']
+	del login_session['picture']
+	del login_session['user_id']
+	del login_session['provider']
+
+	return redirect(url_for('showCategory'))
+
+
 @app.route('/fbconnect', methods=['POST'])
 def fbconnect():
-    if request.args.get('state') != login_session['state']:
-        response = make_response(json.dumps('Invalid state parameter.'), 401)
-        response.headers['Content-Type'] = 'application/json'
-        return response
-    access_token = request.data
-    print ("access token received %s " % access_token)
+	# Validate anti-forgery state token
+	if request.args.get('state') != login_session['state']:
+	    response = make_response(json.dumps('Invalid state parameter.'), 401)
+	    response.headers['Content-Type'] = 'application/json'
+	    return response
 
+	# Gets acces token
+	access_token = request.data
+	print ("access token received %s " % access_token)
 
-    app_id = json.loads(open('fb_client_secrets.json', 'r').read())[
-        'web']['app_id']
-    app_secret = json.loads(
-        open('fb_client_secrets.json', 'r').read())['web']['app_secret']
-    url = 'https://graph.facebook.com/oauth/access_token?grant_type=fb_exchange_token&client_id=%s&client_secret=%s&fb_exchange_token=%s' % (
-        app_id, app_secret, access_token)
-    h = httplib2.Http()
-    result = h.request(url, 'GET')[1]
+	# Gets info from fb clients secrets
+	app_id = json.loads(open('fb_client_secrets.json', 'r').read())['web']['app_id']
+	app_secret = json.loads(open('fb_client_secrets.json', 'r').read())['web']['app_secret']
 
+	url = 'https://graph.facebook.com/oauth/access_token?grant_type=fb_exchange_token&client_id=%s&client_secret=%s&fb_exchange_token=%s' % (app_id, app_secret, access_token)
+	h = httplib2.Http()
+	result = h.request(url, 'GET')[1]
 
-    # Use token to get user info from API
-    userinfo_url = "https://graph.facebook.com/v2.8/me"
-    '''
-    Due to the formatting for the result from the server token exchange we have to
-    split the token first on commas and select the first index which gives us the key : value
-    for the server access token then we split it on colons to pull out the actual token value
-    and replace the remaining quotes with nothing so that it can be used directly in the graph
-    api calls
-    '''
-    token = result.split(',')[0].split(':')[1].replace('"', '')
+	# Use token to get user info from API
+	userinfo_url = "https://graph.facebook.com/v2.4/me"
 
-    url = 'https://graph.facebook.com/v2.8/me?access_token=%s&fields=name,id,email' % token
-    h = httplib2.Http()
-    result = h.request(url, 'GET')[1]
-    # print "url sent for API access:%s"% url
-    # print "API JSON result: %s" % result
-    data = json.loads(result)
-    login_session['provider'] = 'facebook'
-    login_session['username'] = data["name"]
-    login_session['email'] = data["email"]
-    login_session['facebook_id'] = data["id"]
+    # strip expire tag from access token
+	token = result.split("&")[0]
 
-    # The token must be stored in the login_session in order to properly logout
-    login_session['access_token'] = token
+	url = 'https://graph.facebook.com/v2.4/me?%s&fields=name,id,email' % token
+	h = httplib2.Http()
+	result = h.request(url, 'GET')[1]
 
-    # Get user picture
-    url = 'https://graph.facebook.com/v2.8/me/picture?access_token=%s&redirect=0&height=200&width=200' % token
-    h = httplib2.Http()
-    result = h.request(url, 'GET')[1]
-    data = json.loads(result)
+	data = json.loads(result)
+	login_session['provider'] = 'facebook'
+	login_session['username'] = data["name"]
+	login_session['email'] = data["email"]
+	login_session['facebook_id'] = data["id"]
 
-    login_session['picture'] = data["data"]["url"]
+	# Store token in login_session in order to logout
+	stored_token = token.split("=")[1]
+	login_session['access_token'] = stored_token
 
-    # see if user exists
-    user_id = getUserID(login_session['email'])
-    if not user_id:
-        user_id = createUser(login_session)
-    login_session['user_id'] = user_id
+	# Get user picture
+	url = 'https://graph.facebook.com/v2.4/me/picture?%s&redirect=0&height=200&width=200' % token
+	h = httplib2.Http()
+	result = h.request(url, 'GET')[1]
+	data = json.loads(result)
 
-    output = ''
-    output += '<h1>Welcome, '
-    output += login_session['username']
+	login_session['picture'] = data["data"]["url"]
 
-    output += '!</h1>'
-    output += '<img src="'
-    output += login_session['picture']
-    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+	# See if user exists
+	user_id = getUserID(login_session['email'])
+	if not user_id:
+	    user_id = createUser(login_session)
+	login_session['user_id'] = user_id
 
-    flash("Now logged in as %s" % login_session['username'])
-    return output
-
+	return "Login Successful!"
 
 @app.route('/fbdisconnect')
 def fbdisconnect():
-    facebook_id = login_session['facebook_id']
-    # The access token must me included to successfully logout
-    access_token = login_session['access_token']
-    url = 'https://graph.facebook.com/%s/permissions?access_token=%s' % (facebook_id,access_token)
-    h = httplib2.Http()
-    result = h.request(url, 'DELETE')[1]
-    return "you have been logged out"
+	facebook_id = login_session['facebook_id']
+	access_token = login_session['access_token']
 
+	url = 'https://graph.facebook.com/%s/permissions?access_token=%s' % (facebook_id,access_token)
+	h = httplib2.Http()
+	result = h.request(url, 'DELETE')[1]
 
-#########
-#Gconnect#
-##########
+	return "you have been logged out"
 
 @app.route('/gconnect', methods=['POST'])
 def gconnect():
-    # Validate state token
-    if request.args.get('state') != login_session['state']:
-        response = make_response(json.dumps('Invalid state parameter.'), 401)
-        response.headers['Content-Type'] = 'application/json'
-        return response
-    # Obtain authorization code
-    code = request.data
+	# Validate anti-forgery state token
+	if request.args.get('state') != login_session['state']:
+		response = make_response(json.dumps('Invalid state parameter.'), 401)
+		response.headers['Content-Type'] = 'application/json'
+		return response
 
-    try:
-        # Upgrade the authorization code into a credentials object
-        oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
-        oauth_flow.redirect_uri = 'postmessage'
-        credentials = oauth_flow.step2_exchange(code)
-    except FlowExchangeError:
-        response = make_response(
-            json.dumps('Failed to upgrade the authorization code.'), 401)
-        response.headers['Content-Type'] = 'application/json'
-        return response
+	# Obtain authorization code
+	code = request.data
 
-    # Check that the access token is valid.
-    access_token = credentials.access_token
-    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
-           % access_token)
-    h = httplib2.Http()
-    result = json.loads(h.request(url, 'GET')[1])
-    # If there was an error in the access token info, abort.
-    if result.get('error') is not None:
-        response = make_response(json.dumps(result.get('error')), 500)
-        response.headers['Content-Type'] = 'application/json'
-        return response
+	try:
+		# Upgrade the authorization code into a credentials object
+		oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
+		oauth_flow.redirect_uri = 'postmessage'
+		credentials = oauth_flow.step2_exchange(code)
+	except FlowExchangeError:
+		response = make_response(json.dumps('Failed to upgrade the authorization code.'), 401)
+		response.headers['Content-Type'] = 'application/json'
+		return response
 
-    # Verify that the access token is used for the intended user.
-    gplus_id = credentials.id_token['sub']
-    if result['user_id'] != gplus_id:
-        response = make_response(
-            json.dumps("Token's user ID doesn't match given user ID."), 401)
-        response.headers['Content-Type'] = 'application/json'
-        return response
+	# Check that the access token is valid.
+	access_token = credentials.access_token
+	url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s' % access_token)
+	h = httplib2.Http()
+	result = json.loads(h.request(url, 'GET')[1])
 
-    # Verify that the access token is valid for this app.
-    if result['issued_to'] != CLIENT_ID:
-        response = make_response(
-            json.dumps("Token's client ID does not match app's."), 401)
-        print ("Token's client ID does not match app's.")
-        response.headers['Content-Type'] = 'application/json'
-        return response
+	# If there was an error in the access token info, abort.
+	if result.get('error') is not None:
+		response = make_response(json.dumps(result.get('error')), 500)
+		response.headers['Content-Type'] = 'application/json'
+		return response
 
-    stored_access_token = login_session.get('access_token')
-    stored_gplus_id = login_session.get('gplus_id')
-    if stored_access_token is not None and gplus_id == stored_gplus_id:
-        response = make_response(json.dumps('Current user is already connected.'),
-                                 200)
-        response.headers['Content-Type'] = 'application/json'
-        return response
+	# Verify that the access token is used for the intended user.
+	gplus_id = credentials.id_token['sub']
+	if result['user_id'] != gplus_id:
+		response = make_response(json.dumps("Token's user ID doesn't match given user ID."), 401)
+		response.headers['Content-Type'] = 'application/json'
+		return response
 
-    # Store the access token in the session for later use.
-    login_session['access_token'] = credentials.access_token
-    login_session['gplus_id'] = gplus_id
+	# Verify that the access token is valid for this app.
+	if result['issued_to'] != CLIENT_ID:
+		response = make_response(json.dumps("Token's client ID does not match app's."), 401)
+		print ("Token's client ID does not match app's.")
+		response.headers['Content-Type'] = 'application/json'
+		return response
 
-    # Get user info
-    userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
-    params = {'access_token': credentials.access_token, 'alt': 'json'}
-    answer = requests.get(userinfo_url, params=params)
+	stored_access_token = login_session.get('access_token')
+	stored_gplus_id = login_session.get('gplus_id')
 
-    data = answer.json()
+	if stored_access_token is not None and gplus_id == stored_gplus_id:
+		response = make_response(json.dumps('Current user is already connected.'), 200)
+		response.headers['Content-Type'] = 'application/json'
+		return response
 
-    login_session['username'] = data['name']
-    login_session['picture'] = data['picture']
-    login_session['email'] = data['email']
+	# Store the access token in the session for later use.
+	login_session['access_token'] = credentials.access_token
+	login_session['gplus_id'] = gplus_id
 
-    output = ''
-    output += '<h1>Welcome, '
-    output += login_session['username']
-    output += '!</h1>'
-    output += '<img src="'
-    output += login_session['picture']
-    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
-    flash("you are now logged in as %s" % login_session['username'])
-    print ("done!")
-    return output
+	# Get user info
+	userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
+	params = {'access_token': credentials.access_token, 'alt': 'json'}
+	answer = requests.get(userinfo_url, params=params)
 
-    # DISCONNECT - Revoke a current user's token and reset their login_session
+	data = answer.json()
 
+	login_session['username'] = data['name']
+	login_session['picture'] = data['picture']
+	login_session['email'] = data['email']
+	login_session['provider'] = 'google'
+
+	# See if user exists
+	user_id = getUserID(data["email"])
+	if not user_id:
+	    user_id = createUser(login_session)
+	login_session['user_id'] = user_id
+
+	return "Login Successful"
 
 @app.route('/gdisconnect')
 def gdisconnect():
-    access_token = login_session.get('access_token')
-    if access_token is None:
-        print ('Access Token is None')
-        response = make_response(json.dumps('Current user not connected.'), 401)
-        response.headers['Content-Type'] = 'application/json'
-        return response
-    print ('In gdisconnect access token is %s', access_token)
-    print ('User name is: ')
-    print (login_session['username'])
-    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % login_session['access_token']
-    h = httplib2.Http()
-    result = h.request(url, 'GET')[0]
-    print ('result is ')
-    print (result)
-    if result['status'] == '200':
-        del login_session['access_token']
-        del login_session['gplus_id']
-        del login_session['username']
-        del login_session['email']
-        del login_session['picture']
-        response = make_response(json.dumps('Successfully disconnected.'), 200)
-        response.headers['Content-Type'] = 'application/json'
-        return response
-    else:
-        response = make_response(json.dumps('Failed to revoke token for given user.', 400))
-        response.headers['Content-Type'] = 'application/json'
-    return response
+	# Only disconnect a connected user.
+	access_token = login_session.get('access_token')
+
+	if access_token is None:
+		response = make_response(json.dumps('Current user not connected.'), 401)
+		response.headers['Content-Type'] = 'application/json'
+		return response
+
+	url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % access_token
+	h = httplib2.Http()
+	result = h.request(url, 'GET')[0]
+
+	if result['status'] != '200':
+	    # For whatever reason, the given token was invalid.
+	    response = make_response(json.dumps('Failed to revoke token for given user.'), 400)
+	    response.headers['Content-Type'] = 'application/json'
+	    return response
 
 
 ###########
@@ -291,12 +274,17 @@ def CategoryItemsJSON(category_name):
 @app.route('/Catalog/')
 def show_categories():
     category = session.query(Category)
-    return render_template("Categories.html", category=category)
+    # printing Items of Category
+    items = session.query(Items).all()
+    return render_template("Categories.html", category=category,items=items)
 
 
 # new category # Done
 @app.route('/Catalog/new/', methods=['GET', 'POST'])
 def new_category():
+    # Check if user is logged in
+    if 'username' not in login_session:
+        return redirect('/login')
     if request.method == 'POST':
         new_category = Category(name=request.form['name'])
         session.add(new_category)
@@ -308,12 +296,14 @@ def new_category():
 # show Category Items # Done
 @app.route('/Catalog/<path:category_name>/')
 def showCategory(category_name):
-    # login = True
-    # if login==True:
-    # edit_category(category_name)
+
+
+
     category = session.query(Category).filter_by(name=category_name).one()
     items = session.query(Items).filter_by(category=category)
-    return render_template("items.html", Category=category, Items=items)
+    # Get creator of item
+    creator = getUserInfo(category.user_id)
+    return render_template("items.html", Category=category, Items=items ,creator =creator)
 
 
 # edit category items
@@ -324,13 +314,18 @@ def showCategory(category_name):
 def show_item( category_name ,item_name):
     category = session.query(Category).filter_by(name=category_name).one()
     items = session.query(Items).filter_by(name=item_name).one()
-    return render_template("item.html", item=items, category=category.name)
+    # Get creator of item
+    creator = getUserInfo(items.user_id)
+    return render_template("item.html", item=items, category=category.name, creator=creator)
 
 
 # add new items to the category
 
 @app.route('/Catalog/<path:category_name>/new/', methods=['GET', 'POST'])
 def new_items(category_name):
+    # Check if user is logged in
+    if 'username' not in login_session:
+        return redirect('/login')
     category = session.query(Category).filter_by(name=category_name).one()
     print(category.name)
     if request.method == 'POST':
@@ -348,7 +343,17 @@ def new_items(category_name):
 
 @app.route('/Catalog/<path:category_name>/<path:item_name>/edit/', methods=['GET', 'POST'])
 def edit_item(category_name, item_name):
+    # Check if user is logged in
+    if 'username' not in login_session:
+        return redirect('/login')
     item = session.query(Items).filter_by(name=item_name).one()
+    # Get creator of item
+    creator = getUserInfo(item.user_id)
+
+    # Check if logged in user is creator of category item
+    if creator.id != login_session['user_id']:
+
+        return redirect('/login')
     category = session.query(Category).filter_by(name=category_name).one()
     if request.method == 'POST':
         if request.form['name']:
@@ -367,7 +372,16 @@ def edit_item(category_name, item_name):
 
 @app.route('/Catalog/<path:category_name>/<path:item_name>/delete/')
 def delete_item(category_name, item_name):
+    # Check if user is logged in
+    if 'username' not in login_session:
+        return redirect('/login')
     item = session.query(Items).filter_by(name=item_name).one()
+    # Get creator of item
+    creator = getUserInfo(item.user_id)
+
+    # Check if logged in user is creator of category item
+    if creator.id != login_session['user_id']:
+       return redirect('/login')
     category = session.query(Category).filter_by(name=category_name).one()
     if request.method == 'POST':
         session.delete(item)
@@ -384,6 +398,9 @@ def delete_item(category_name, item_name):
 
 @app.route('/Catalog/<path:category_name>/delete/')
 def delete_Category(category_name):
+    # Check if user is logged in
+    if 'username' not in login_session:
+        return redirect('/login')
     #item = session.query(Items).filter_by(name=item_name).one()
     category = session.query(Category).filter_by(name=category_name).one()
     if request.method == 'POST':
@@ -398,7 +415,7 @@ def delete_Category(category_name):
 if __name__ == '__main__':
     app.secret_key = 'super_secret_key'
     app.debug = True
-    #app.run(host='0.0.0.0', port=5000)
-    app.run(debug=True)
+    app.run(port=5000, host='localhost')
+
